face_detection/requirements.txt
click==8.1.3
cmake==3.24.2
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
numpy==1.23.4
opencv-python==4.6.0.66
paho-mqtt==1.6.1
Pillow==9.3.0
state-machine==0.2.10
